Fix REAL testing against float64 instead of float32

+++ tests/libtest/samples.go
@@ -92,9 +92,9 @@ func compareDecimal(recv, expect *types.Decimal) bool {
 // TODO: -null database/sql.NullFloat64
 var samplesFloat = []float64{math.SmallestNonzeroFloat64, math.MaxFloat64}
 
-//go:generate go run ./gen_type.go Real float64
-// TODO: -null database/sql.NullFloat64
-var samplesReal = []float64{
+//go:generate go run ./gen_type.go Real float32
+// TODO: -null database/sql.NullFloat32
+var samplesReal = []float32{
 	math.SmallestNonzeroFloat32,
 	math.MaxFloat32,
 }
